requirements.txt
moviepy
nltk
streamlit
SpeechRecognition
transformers
pyttsx3
pyaudio
torch
pydub
opencv-python
streamlit
google-generativeai
python-dotenv
gtts